Splited npc script to stave of ancients script
  * Move reveal form to npcstaveqai
  * Change enum constant names
  * Refactor DoAction
  * Refactor feign death reset
