+++ src/server/scripts/Kalimdor/zone_winterspring.cpp
@@ -17,7 +17,6 @@ npc_ranshalla
 go_elune_fire
 EndContentData */
 
-#include "CreatureTextMgr.h"
 #include "Player.h"
 #include "ScriptedCreature.h"
 #include "ScriptedEscortAI.h"
@@ -53,309 +52,6 @@ enum eStaveQuest
 
     SPELL_FOOLS_PLIGHT                      = 23504,
     SPELL_SOUL_FLAME                        = 23272,
-};
-
-enum NPCArtorius
-{
-  ARTORIUS_EVENT_ENCOUNTER_START            = 1,
-  ARTORIUS_EVENT_REVEAL                     = 2,
-  ARTORIUS_EVENT_RANGE_CHECK                = 3,
-  ARTORIUS_EVENT_UNFAIR_FIGHT               = 4,
-  ARTORIUS_EVENT_DEMONIC_DOOM               = 5,
-
-  ARTORIUS_GOSSIP_TEXT                      = 7045,   // npc_text.ID               "How wonderful to see another person in..."
-  ARTORIUS_GOSSIP_OPTION_TEXT               = 14531,  // creature_text.CreatureID  "I know you as Artorius the Doombringer..."
-  ARTORIUS_SAY                              = 9787,   // broadcast_text.ID         "Your soul is mine, weakling."
-
-  ARTORIUS_NORMAL_ENTRY                     = 14531,  // creature_template.entry
-  ARTORIUS_EVIL_ENTRY                       = 14535,  // creature_template.entry
-
-  ARTORIUS_HEAD                             = 18955,   // item_template.entry
-
-  ARTORIUS_SPELL_DEMONIC_DOOM               = 23298,
-  ARTORIUS_SPELL_STINGING_TRAUMA            = 23299
-};
-
-struct NPCStaveQAI
-{
-    ObjectGuid pGUID;
-    ThreatContainer::StorageType threatList;
-
-    void StorePlayerGUID()
-    {
-        if (!pGUID.IsEmpty())
-        {
-            return;
-        }
-
-        for (ThreatContainer::StorageType::const_iterator itr = threatList.begin(); itr != threatList.end(); ++itr)
-        {
-            if ((*itr)->getTarget()->GetTypeId() == TYPEID_PLAYER)
-            {
-                pGUID = (*itr)->getUnitGuid();
-            }
-        }
-    }
-
-    void InitMembers(Creature* creature)
-    {
-        threatList = creature->getThreatManager().getThreatList();
-        StorePlayerGUID();
-    }
-
-    bool IsAllowedEntry(uint32 entry)
-    {
-        uint32 allowedEntries[4] = { 0, 12999, 19833, 19921 }; //player, World Invisible Trigger(traps) and snake trap snakes
-        bool isAllowed = std::find(std::begin(allowedEntries), std::end(allowedEntries), entry) != std::end(allowedEntries);
-        return isAllowed;
-    }
-
-    bool IsFairFight(Creature* creature)
-    {
-        for (ThreatContainer::StorageType::const_iterator itr = threatList.begin(); itr != threatList.end(); ++itr)
-        {
-            Unit* unit = ObjectAccessor::GetUnit(*creature, (*itr)->getUnitGuid());
-            bool allowedEntry = IsAllowedEntry(unit->GetEntry());
-
-            if (!(*itr)->getThreat())
-            {
-                // if target threat is 0 its fair, this prevents despawn in the case when
-                // there is a bystander since UpdateVictim adds nearby enemys to the threatlist
-                continue;
-            }
-
-            if (unit->IsPlayer())
-            {
-                if (pGUID != unit->GetGUID())
-                {
-                    // if there is another player in the threatlist its unfair
-                    return false;
-                }
-            }
-            else
-            {
-                if (!pGUID.IsEmpty() && unit->GetOwnerGUID() != pGUID)
-                {
-                    // if a creature attacking isn't owned by the player its unfair
-                    return false;
-                }
-                else if (!allowedEntry)
-                {
-                    // if not in the whitelist its unfair
-                    return false;
-                }
-            }
-        }
-
-        return true;
-    }
-
-    bool ValidThreatlist(Creature* creature)
-    {
-        if (threatList.size() == 1)
-        {
-            return true;
-        }
-
-        bool isFair = IsFairFight(creature);
-
-        return isFair;
-    }
-
-    void SetHomePosition(Creature* creature)
-    {
-        Position homePosition;
-        creature->GetPosition(&homePosition);
-
-        if (homePosition.IsPositionValid())
-        {
-            creature->SetHomePosition(homePosition);
-            creature->SetWalk(false); // maybe we dont need this
-        }
-    }
-};
-
-class npc_artorius : public CreatureScript
-{
-public:
-    npc_artorius() : CreatureScript("npc_artorius") { }
-
-    CreatureAI* GetAI(Creature* creature) const override
-    {
-        return new npc_artoriusAI(creature);
-    }
-
-    struct npc_artoriusAI : public ScriptedAI, public NPCStaveQAI
-    {
-        npc_artoriusAI(Creature *creature) : ScriptedAI(creature) { }
-
-        EventMap events;
-        Position eventPosition;
-        bool encounterStarted;
-
-        void Reset() override
-        {
-            encounterStarted = false;
-            pGUID.Clear();
-            events.Reset();
-
-            if (me->HasAura(ARTORIUS_SPELL_STINGING_TRAUMA))
-            {
-                me->RemoveAura(ARTORIUS_SPELL_STINGING_TRAUMA);
-            }
-        }
-
-        void AttackStart(Unit* target) override
-        {
-            RevealForm();
-
-            if (me->GetEntry() == ARTORIUS_EVIL_ENTRY)
-            {
-                InitMembers(me);
-                events.ScheduleEvent(ARTORIUS_EVENT_DEMONIC_DOOM, urand(3000, 5000));
-                events.ScheduleEvent(ARTORIUS_EVENT_RANGE_CHECK, 1000);
-                events.ScheduleEvent(ARTORIUS_EVENT_UNFAIR_FIGHT, 1000);
-            }
-
-            ScriptedAI::AttackStart(target);
-        }
-
-        void RevealForm()
-        {
-            if (encounterStarted && me->GetEntry() == ARTORIUS_NORMAL_ENTRY)
-            {
-                me->UpdateEntry(ARTORIUS_EVIL_ENTRY);
-                me->DespawnOrUnsummon(900000);
-            }
-        }
-
-        void UpdateAI(uint32 diff) override
-        {
-            events.Update(diff);
-            uint32 eventId = events.ExecuteEvent();
-
-            switch (eventId)
-            {
-                case ARTORIUS_EVENT_ENCOUNTER_START:
-                    me->MonsterSay(ARTORIUS_SAY, LANG_UNIVERSAL, 0);
-                    me->RemoveFlag(UNIT_FIELD_FLAGS, UNIT_FLAG_NON_ATTACKABLE);
-                    events.ScheduleEvent(ARTORIUS_EVENT_REVEAL, 8000);
-                    break;
-                case ARTORIUS_EVENT_REVEAL:
-                    RevealForm();
-                    break;
-            }
-
-
-            if (UpdateVictim())
-            {
-                // This should prevent hunters from staying in combat when feign death is used and their pet has 0 threat
-                if (!pGUID.IsEmpty()
-                  && me->GetVictim()->GetOwnerGUID() == pGUID
-                  && ObjectAccessor::GetPlayer(*me, pGUID)->HasAura(5384)
-                  && !IsAllowedEntry(me->GetVictim()->GetEntry())
-                )
-                {
-                    pGUID.Clear();
-                    EnterEvadeMode();
-                    return;
-                }
-            }
-            else
-            {
-                return;
-            }
-
-            if (me->HasUnitState(UNIT_STATE_CASTING))
-            {
-                return;
-            }
-
-            switch (eventId)
-            {
-                case ARTORIUS_EVENT_RANGE_CHECK:
-                    if (!me->GetVictim()->IsWithinDist2d(me, 60.0f))
-                    {
-                        EnterEvadeMode();
-                    }
-                    else
-                    {
-                        events.RepeatEvent(2000);
-                    }
-                    break;
-                case ARTORIUS_EVENT_UNFAIR_FIGHT:
-                    if (!ValidThreatlist(me))
-                    {
-                        SetHomePosition(me);
-                        me->SetFlag(UNIT_FIELD_FLAGS, UNIT_FLAG_DISABLE_MOVE | UNIT_FLAG_NON_ATTACKABLE | UNIT_FLAG_NOT_ATTACKABLE_1 | UNIT_FLAG_IMMUNE_TO_PC | UNIT_FLAG_IMMUNE_TO_NPC);
-                        me->DespawnOrUnsummon(5000);
-                        break;
-                    }
-                    events.RepeatEvent(2000);
-                    break;
-                case ARTORIUS_EVENT_DEMONIC_DOOM:
-                    if (!me->GetVictim()->HasAura(ARTORIUS_SPELL_DEMONIC_DOOM))
-                    {
-                        me->CastSpell(me->GetVictim(), ARTORIUS_SPELL_DEMONIC_DOOM, false);
-                    }
-                    events.RepeatEvent(urand(5000, 10000));
-                    break;
-            }
-
-            DoMeleeAttackIfReady();
-        }
-
-        void SpellHit(Unit* /*Caster*/, const SpellInfo* Spell) override
-        {
-            uint32 serpentStings[12] = { 1978, 13549, 13550, 13551, 13552, 13553, 13554, 13555, 25295, 27016, 49000, 49001 };
-
-            if (me->GetEntry() == ARTORIUS_EVIL_ENTRY)
-            {
-                bool applyAura = std::find(std::begin(serpentStings), std::end(serpentStings), Spell->Id) != std::end(serpentStings);
-
-                if (applyAura)
-                {
-                    me->AddAura(ARTORIUS_SPELL_STINGING_TRAUMA, me);
-                }
-            }
-        }
-
-        void DoAction(int32 action) override
-        {
-            if (action == ARTORIUS_EVENT_ENCOUNTER_START)
-            {
-                encounterStarted = true;
-                me->GetMotionMaster()->MoveIdle();
-                me->GetMotionMaster()->Clear();
-                SetHomePosition(me);
-
-                me->SetFlag(UNIT_FIELD_FLAGS, UNIT_FLAG_NON_ATTACKABLE);
-                me->RemoveFlag(UNIT_NPC_FLAGS, UNIT_NPC_FLAG_GOSSIP);
-                events.ScheduleEvent(ARTORIUS_EVENT_ENCOUNTER_START, 2000);
-            }
-        }
-    };
-
-    bool OnGossipSelect(Player* player, Creature* creature, uint32 /*sender*/, uint32 /*action*/) override
-    {
-        CloseGossipMenuFor(player);
-        creature->AI()->DoAction(ARTORIUS_EVENT_ENCOUNTER_START);
-
-        return true;
-    }
-
-    bool OnGossipHello(Player *player, Creature * creature) override
-    {
-        if (player->GetQuestStatus(QUEST_STAVE_OF_THE_ANCIENTS) == QUEST_STATUS_INCOMPLETE && !player->HasItemCount(ARTORIUS_HEAD, 1, true))
-        {
-            std::string const& gossipOptionText = sCreatureTextMgr->GetLocalizedChatString(ARTORIUS_GOSSIP_OPTION_TEXT, 0, 0, 0, LOCALE_enUS);
-            AddGossipItemFor(player, GOSSIP_ICON_CHAT, gossipOptionText, GOSSIP_SENDER_MAIN, 0);
-        }
-
-        SendGossipMenuFor(player, ARTORIUS_GOSSIP_TEXT, creature->GetGUID());
-
-        return true;
-    }
 };
 
 class npc_stave_of_the_ancients : public CreatureScript
@@ -1098,7 +794,6 @@ public:
 void AddSC_winterspring()
 {
     // Ours
-    new npc_artorius();
     new npc_stave_of_the_ancients();
 
     // Theirs
