+++ src/server/scripts/World/world_script_loader.cpp
@@ -19,6 +19,7 @@ void AddSC_chat_log(); // location: scripts\World\chat_log.cpp
 void AddSC_character_creation();
 void AddSC_action_ip_logger(); // location: scripts\World\action_ip_logger.cpp
 void AddSC_player_scripts();
+void AddSC_npc_stave_of_ancients();
 
 // The name of this function should match:
 // void Add${NameOfDirectory}Scripts()
@@ -39,4 +40,5 @@ void AddWorldScripts()
     AddSC_character_creation();
     AddSC_action_ip_logger(); // location: scripts\World\action_ip_logger.cpp
     AddSC_player_scripts();
+    AddSC_npc_stave_of_ancients();
 }
